+++ packages/client/hooks/useMenu.ts
@@ -51,7 +51,7 @@ const useMenu = <T extends HTMLElement = HTMLButtonElement>(
     if (options.loadingWidth) return options.loadingWidth
     const bbox = getBBox(originRef.current)
     return Math.max(40, bbox ? bbox.width : 40)
-  }, [originRef.current /* eslint-disable-line react-hooks/exhaustive-deps */])
+  }, [originRef.current])
   const {loadingDelay, loadingDelayRef} = useLoadingDelay()
   const menuPortal = useMenuPortal(
     portal,