+++ packages/client/modules/userDashboard/components/OrgUserRow/OrgMemberRow.tsx
@@ -1,9 +1,8 @@
 import styled from '@emotion/styled'
 import graphql from 'babel-plugin-relay/macro'
 import {format} from 'date-fns'
-import React, {forwardRef, Ref} from 'react'
+import React from 'react'
 import {useFragment} from 'react-relay'
-import useAtmosphere from '~/hooks/useAtmosphere'
 import {
   OrgMemberRow_organization$data,
   OrgMemberRow_organization$key
@@ -13,8 +12,6 @@ import {
   OrgMemberRow_organizationUser$key
 } from '../../../../__generated__/OrgMemberRow_organizationUser.graphql'
 import Avatar from '../../../../components/Avatar/Avatar'
-import FlatButton, {FlatButtonProps} from '../../../../components/FlatButton'
-import IconLabel from '../../../../components/IconLabel'
 import RowActions from '../../../../components/Row/RowActions'
 import RowInfo from '../../../../components/Row/RowInfo'
 import RowInfoHeader from '../../../../components/Row/RowInfoHeader'
@@ -23,47 +20,27 @@ import RowInfoLink from '../../../../components/Row/RowInfoLink'
 import BaseTag from '../../../../components/Tag/BaseTag'
 import InactiveTag from '../../../../components/Tag/InactiveTag'
 import RoleTag from '../../../../components/Tag/RoleTag'
-import {MenuPosition} from '../../../../hooks/useCoords'
-import useMenu from '../../../../hooks/useMenu'
 import useModal from '../../../../hooks/useModal'
 import defaultUserAvatar from '../../../../styles/theme/images/avatar-user.svg'
 import lazyPreload from '../../../../utils/lazyPreload'
-import withMutationProps, {WithMutationProps} from '../../../../utils/relay/withMutationProps'
 
-interface Props extends WithMutationProps {
+const ActionsBlock = styled('div')({
+  alignItems: 'center',
+  display: 'flex',
+  justifyContent: 'flex-end'
+})
+
+interface Props {
   billingLeaderCount: number
   orgAdminCount: number
   organizationUser: OrgMemberRow_organizationUser$key
   organization: OrgMemberRow_organization$key
 }
 
-const StyledButton = styled(FlatButton)({
-  paddingLeft: 0,
-  paddingRight: 0,
-  width: '100%'
-})
-
-const StyledFlatButton = styled(FlatButton)({
-  paddingLeft: 16,
-  paddingRight: 16
-})
-
-const MenuButton = forwardRef((props: FlatButtonProps, ref: Ref<HTMLButtonElement>) => (
-  <StyledButton {...props} disabled={props.disabled} ref={ref}>
-    <IconLabel icon='more_vert' />
-  </StyledButton>
-))
-
 const LeaveOrgModal = lazyPreload(
   () => import(/* webpackChunkName: 'LeaveOrgModal' */ '../LeaveOrgModal/LeaveOrgModal')
 )
 
-const BillingLeaderActionMenu = lazyPreload(
-  () =>
-    import(
-      /* webpackChunkName: 'BillingLeaderActionMenu' */ '../../../../components/BillingLeaderActionMenu'
-    )
-)
 const OrgAdminActionMenu = lazyPreload(
   () =>
     import(/* webpackChunkName: 'OrgAdminActionMenu' */ '../../../../components/OrgAdminActionMenu')
@@ -117,95 +94,47 @@ const UserInfo: React.FC<UserInfoProps> = ({
 )
 
 interface UserActionsProps {
-  isViewerOrgAdmin: boolean
-  isViewerBillingLeader: boolean
-  isViewerLastOrgAdmin: boolean
-  isViewerLastBillingLeader: boolean
   organization: OrgMemberRow_organization$data
   organizationUser: OrgMemberRow_organizationUser$data
   preferredName: string
-  viewerId: string
 }
 
 const UserActions: React.FC<UserActionsProps> = ({
-  isViewerOrgAdmin,
-  isViewerBillingLeader,
-  isViewerLastOrgAdmin,
-  isViewerLastBillingLeader,
   organizationUser,
   organization,
-  preferredName,
-  viewerId
+  preferredName
 }) => {
-  const {orgId} = organization
+  const {id: orgId} = organization
   const {
-    user: {userId}
+    user: {id: userId}
   } = organizationUser
-  const {togglePortal, originRef, menuPortal, menuProps} = useMenu(MenuPosition.UPPER_RIGHT)
   const {togglePortal: toggleLeave, modalPortal: leaveModal} = useModal()
   const {togglePortal: toggleRemove, modalPortal: removeModal} = useModal()
-  const actionMenuProps = {
-    menuProps,
-    originRef,
-    togglePortal,
-    toggleLeave,
-    toggleRemove,
-    isViewerLastOrgAdmin,
-    isViewerLastBillingLeader,
-    organization,
-    organizationUser
-  }
-
-  const showLeaveButton = !isViewerOrgAdmin && !isViewerBillingLeader && viewerId === userId
-
   return (
     <RowActions>
-      <div className='flex items-center justify-end'>
-        {showLeaveButton && (
-          <StyledFlatButton onClick={toggleLeave} onMouseEnter={LeaveOrgModal.preload}>
-            Leave Organization
-          </StyledFlatButton>
-        )}
-        {(isViewerOrgAdmin || (isViewerBillingLeader && !isViewerLastBillingLeader)) && (
-          <div className='ml-2 w-8'>
-            <MenuButton
-              onClick={togglePortal}
-              onMouseEnter={
-                isViewerOrgAdmin ? OrgAdminActionMenu.preload : BillingLeaderActionMenu.preload
-              }
-              ref={originRef}
-            />
-          </div>
-        )}
-        {isViewerOrgAdmin && menuPortal(<OrgAdminActionMenu {...actionMenuProps} />)}
-        {!isViewerOrgAdmin &&
-          isViewerBillingLeader &&
-          menuPortal(<BillingLeaderActionMenu {...actionMenuProps} />)}
+      <ActionsBlock>
+        <OrgAdminActionMenu
+          organization={organization}
+          organizationUser={organizationUser}
+          toggleLeave={toggleLeave}
+          toggleRemove={toggleRemove}
+        />
         {leaveModal(<LeaveOrgModal orgId={orgId} />)}
         {removeModal(
           <RemoveFromOrgModal orgId={orgId} userId={userId} preferredName={preferredName} />
         )}
-      </div>
+      </ActionsBlock>
     </RowActions>
   )
 }
 
 const OrgMemberRow = (props: Props) => {
-  const atmosphere = useAtmosphere()
-  const {
-    billingLeaderCount,
-    orgAdminCount,
-    organizationUser: organizationUserRef,
-    organization: organizationRef
-  } = props
+  const {organizationUser: organizationUserRef, organization: organizationRef} = props
 
   const organization = useFragment(
     graphql`
       fragment OrgMemberRow_organization on Organization {
-        isViewerBillingLeader: isBillingLeader
-        isViewerOrgAdmin: isOrgAdmin
-        orgId: id
-        ...BillingLeaderActionMenu_organization
+        id
         ...OrgAdminActionMenu_organization
       }
     `,
@@ -216,7 +145,7 @@ const OrgMemberRow = (props: Props) => {
     graphql`
       fragment OrgMemberRow_organizationUser on OrganizationUser {
         user {
-          userId: id
+          id
           email
           inactive
           picture
@@ -224,27 +153,19 @@ const OrgMemberRow = (props: Props) => {
           lastSeenAt
         }
         role
-        ...BillingLeaderActionMenu_organizationUser
         ...OrgAdminActionMenu_organizationUser
       }
     `,
     organizationUserRef
   )
 
-  const {isViewerBillingLeader, isViewerOrgAdmin} = organization
-
   const {
     user: {email, inactive, picture, preferredName, lastSeenAt},
     role
   } = organizationUser
 
-  const {viewerId} = atmosphere
-
   const isBillingLeader = role === 'BILLING_LEADER'
   const isOrgAdmin = role === 'ORG_ADMIN'
-  const isViewerLastBillingLeader =
-    isViewerBillingLeader && isBillingLeader && billingLeaderCount === 1
-  const isViewerLastOrgAdmin = isViewerOrgAdmin && isOrgAdmin && orgAdminCount === 1
   const formattedLastSeenAt = lastSeenAt ? format(new Date(lastSeenAt), 'yyyy-MM-dd') : 'Never'
 
   return (
@@ -268,18 +189,13 @@ const OrgMemberRow = (props: Props) => {
       </td>
       <td className='w-1/5 py-3 px-2 align-middle'>
         <UserActions
-          isViewerOrgAdmin={isViewerOrgAdmin}
-          isViewerBillingLeader={isViewerBillingLeader}
-          isViewerLastOrgAdmin={isViewerLastOrgAdmin}
-          isViewerLastBillingLeader={isViewerLastBillingLeader}
           organizationUser={organizationUser}
           organization={organization}
           preferredName={preferredName}
-          viewerId={viewerId}
         />
       </td>
     </tr>
   )
 }
 
-export default withMutationProps(OrgMemberRow)
+export default OrgMemberRow