+++ packages/client/mutations/ArchiveTeamMutation.ts
@@ -106,6 +106,7 @@ export const archiveTeamTeamUpdater: SharedUpdater<ArchiveTeamMutation_team$data
   const orgs = viewer.getLinkedRecords('organizations')!
   orgs.forEach((org) => {
     safeRemoveNodeFromArray(teamId, org, 'teams')
+    safeRemoveNodeFromArray(teamId, org, 'allTeams')
   })
 
   const notification = payload.getLinkedRecord('notification')