+++ packages/server/dataloader/foreignKeyLoaderMakers.ts
@@ -2,8 +2,11 @@ import getKysely from '../postgres/getKysely'
 import {getTeamPromptResponsesByMeetingIds} from '../postgres/queries/getTeamPromptResponsesByMeetingIds'
 import {
   selectAgendaItems,
+  selectComments,
   selectOrganizations,
   selectRetroReflections,
+  selectSlackAuths,
+  selectSlackNotifications,
   selectSuggestedAction,
   selectTeams,
   selectTemplateDimension,
@@ -189,5 +192,26 @@ export const agendaItemsByMeetingId = foreignKeyLoaderMaker(
   'meetingId',
   async (meetingIds) => {
     return selectAgendaItems().where('meetingId', 'in', meetingIds).orderBy('sortOrder').execute()
+  }
+)
+
+export const slackAuthByUserId = foreignKeyLoaderMaker('slackAuths', 'userId', async (userIds) => {
+  return selectSlackAuths().where('userId', 'in', userIds).execute()
+})
+
+export const slackNotificationsByTeamId = foreignKeyLoaderMaker(
+  'slackNotifications',
+  'teamId',
+  async (teamIds) => {
+    return selectSlackNotifications().where('teamId', 'in', teamIds).execute()
+  }
+)
+
+export const _pgcommentsByDiscussionId = foreignKeyLoaderMaker(
+  '_pgcomments',
+  'discussionId',
+  async (discussionIds) => {
+    // include deleted comments so we can replace them with tombstones
+    return selectComments().where('discussionId', 'in', discussionIds).execute()
   }
 )